+++ crates/ryo3/src/libs.rs
@@ -30,6 +30,8 @@ pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {
     ryo3_jiff::pymod_add(m)?;
     #[cfg(feature = "jiter")]
     ryo3_jiter::pymod_add(m)?;
+    #[cfg(feature = "lz4rip")]
+    ryo3_lz4rip::pymod_add(m)?;
     #[cfg(feature = "memchr")]
     ryo3_memchr::pymod_add(m)?;
     #[cfg(feature = "regex")]